#17 followed token format with experience

+++ jawbs-server/server/controllers/experience.ts
@@ -1,14 +1,17 @@
-import { Request, Response } from "express";
+import { Request, Response, NextFunction } from "express";
 import prisma from "../libs/prisma.js";
 import { Address } from "../types/index.js";
+import ClientError from "../libs/client-error.js";
 
-const create = async (req: Request, res: Response) => {
-  const { experiences, userId }: any = req.body;
-  const experienceList = [];
+const create = async (req: Request, res: Response, next: NextFunction) => {
+  const userId = req.user ? req.user.id : undefined;
+
   if (!userId) {
-    res.status(400).json("Needs an authorized user");
-    return;
+    throw new ClientError(400, "invalid credentials");
   }
+
+  const { experiences }: any = req.body;
+
   try {
     for (const experience of experiences) {
       const { address } = experience;
@@ -37,31 +40,33 @@ const create = async (req: Request, res: Response) => {
         userId,
         address: createAddress,
       };
+
       if (!hasAddress) {
         delete experienceFormData.address;
       }
 
       // eslint-disable-next-line
-      const newExperiences = prisma.experience.create({
+      await prisma.experience.create({
         data: experienceFormData,
       });
-      experienceList.push(newExperiences);
     }
 
-    res.status(201).json({
-      data: experienceList,
-      message: "Created experiences",
-    });
+    res.status(201).json({ message: "created experiences" });
   } catch (err) {
-    res.status(400).json({ message: err });
+    next(err);
   } finally {
     prisma.$disconnect();
   }
 };
 
-const read = async (req: Request, res: Response) => {
+const read = async (req: Request, res: Response, next: NextFunction) => {
+  const id = req.user ? req.user.id : undefined;
+
+  if (!id) {
+    throw new ClientError(400, "invalid credentials");
+  }
+
   try {
-    const { id } = req.params;
     const experience = await prisma.experience.findUnique({
       where: {
         id,
@@ -69,15 +74,19 @@ const read = async (req: Request, res: Response) => {
     });
     res.status(200).json(experience);
   } catch (err) {
-    res.status(400).json({ message: err });
+    next(err);
   } finally {
     await prisma.$disconnect();
   }
 };
 
-const update = async (req: Request, res: Response) => {
+const update = async (req: Request, res: Response, next: NextFunction) => {
+  const id = req.user ? req.user.id : undefined;
+  if (!id) {
+    throw new ClientError(400, "invalid credentials");
+  }
+
   try {
-    const { id } = req.params;
     const updates = await req.body;
     const updatedUser = await prisma.user.update({
       where: { id },
@@ -85,21 +94,26 @@ const update = async (req: Request, res: Response) => {
     });
     res.status(200).json(updatedUser);
   } catch (err) {
-    res.status(400).json({ message: err });
+    next(err);
   } finally {
     await prisma.$disconnect();
   }
 };
 
-const destroy = async (req: Request, res: Response) => {
+const destroy = async (req: Request, res: Response, next: NextFunction) => {
+  const id = req.user ? req.user.id : undefined;
+
+  if (!id) {
+    throw new ClientError(400, "invalid credentials");
+  }
+
   try {
-    const { id } = req.params;
     const deletedUser = await prisma.experience.delete({
       where: { id },
     });
     res.status(200).json(deletedUser);
   } catch (err) {
-    res.status(400).json({ message: err });
+    next(err);
   } finally {
     await prisma.$disconnect();
   }
